+++ src/sites/link/1ink.cc.js
@@ -1,0 +1,14 @@
+_.register({
+  rule: {
+    host: /^1ink\.cc$/,
+    path: /^\/\w+$/,
+  },
+  async ready () {
+    let url = document.head.querySelector('[name=keywords]').content;
+    const urlCheck = url.match(/^https?:\/\//);
+    if (!urlCheck) {
+      url = 'http://' + url;
+    }
+    await $.openLink(url);
+  },
+});