+++ src/sites/link/bc.vc.js
@@ -76,7 +76,8 @@
   _.register({
     rule: {
       host: [
-        /^(1tk|hit|adbla|tl7|mylink)\.us$/,
+        /^(1tk|hit|adbla|tl7)\.us$/,
+        /^mylink\.(us|zone)$/,
         /^gx\.si$/,
         /^adwat\.ch$/,
         /^(fly2url|urlwiz|xafox)\.com$/,