+++ src/sites/link/coeg.in.js
@@ -5,7 +5,8 @@
       host: [
         /^link\.animagz\.org$/,
         /^(coeg|disingkat|gunting)\.in$/,
-        /^www\.telondasmu\.com$/,
+        /^www\.(telondasmu|siotong|siherp)\.com$/,
+        /^www\.greget\.space$/,
       ],
       path: /^\/\w+$/,
     },
@@ -50,6 +51,9 @@
       };
     case 'coeg.in':
     case 'www.telondasmu.com':
+    case 'www.siotong.com':
+    case 'www.siherp.com':
+    case 'www.greget.space':
       return () => {
         const a = $('.download-link a');
         return a.href;